requirements.txt
openai>=1.12.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
flask>=3.0.0
flask-cors>=4.0.0
ezdxf>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
# Mayo viewer dependencies
open3d>=0.17.0
trimesh>=4.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
numpy>=1.24.0
